// src/app/business/dashboard-z1/interfaces/dashboard-indicador.interface.ts
export interface DashboardIndicador {
    indId: string;
    indNombre: string;
    indUnidad: string;
    indTipo: string;
    indOptimo: string | null;
    indAtencion: string | null;
    indCuidado: string | null;
    indNombreOriginal: string;
    indOrden: number;
    indUbicacion: string;
    indOrigen: string | null;
    indCampo: string | null;
}