// src/app/business/dashboard-z1/interfaces/dashboard-lugar.interface.ts
export interface DashboardLugar {
    parroquiaId: string;
    // provinciaId:     number;
    // cantonId:        number;
    provinciaNombre: string;
    cantonNombre: string;
    parroquiaNombre: string;
    provinciaDpa: string;
    cantonDpa: string;
    parroquiaDpa: string;
    // geojson:         null;
    estado: string | null;
}